requirements.txt
alembic==1.7.6
boto3==1.21.13
botocore==1.24.13
click==8.0.4
Flask==2.0.3
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
Mako==1.1.6
MarkupSafe==2.1.0
pdfkit==1.0.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.2
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
Werkzeug==2.0.3
